fix: redrive script scaffolding
