+++ scripts/aws/aws-redrive-sqs/package.json
@@ -1,19 +1,20 @@
 {
-  "name": "@go-automation/aws-redrive-sqs",
+  "name": "aws-redrive-sqs",
   "version": "1.0.0",
   "private": true,
   "type": "module",
   "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "tsx src/index.ts"
+    "start": "pnpm build && node dist/index.js",
+    "dev": "tsx src/index.ts",
+    "watch": "tsc --watch",
+    "clean": "rm -rf dist *.tsbuildinfo"
   },
   "devDependencies": {
-    "@types/node": "^22.13.0"
+    "@types/node": "catalog:"
   },
   "dependencies": {
-    "@go-automation/go-common": "workspace:*",
-    "@aws-sdk/client-sqs": "^3.750.0"
+    "@go-automation/go-common": "workspace:*"
   }
 }
