+++ scripts/aws/aws-redrive-sqs/src/types/AwsRedriveSqsConfig.ts
@@ -1,0 +1,25 @@
+/**
+ * Configuration for the AWS Redrive SQS script.
+ */
+export interface AwsRedriveSqsConfig {
+  /** AWS SSO profile name */
+  readonly awsProfile: string;
+
+  /** Source SQS queue name or URL */
+  readonly sourceQueue: string;
+
+  /** Target SQS queue name or URL */
+  readonly targetQueue: string;
+
+  /** Maximum number of messages to move */
+  readonly limit?: number;
+
+  /** Simulate the move without actual sending/deleting */
+  readonly dryRun: boolean;
+
+  /** Visibility timeout for received messages (seconds) */
+  readonly visibilityTimeout: number;
+
+  /** Batch size for SQS operations (1-10) */
+  readonly batchSize: number;
+}
