+++ scripts/aws/aws-redrive-sqs/src/types/index.ts
@@ -1,9 +1,1 @@
-export interface AwsRedriveSqsConfig {
-  readonly awsProfile: string;
-  readonly sourceQueue: string;
-  readonly targetQueue: string;
-  readonly limit?: number;
-  readonly dryRun: boolean;
-  readonly visibilityTimeout: number;
-  readonly batchSize: number;
-}
+export * from './AwsRedriveSqsConfig.js';
