+++ scripts/aws/aws-redrive-sqs/tsconfig.json
@@ -1,9 +1,16 @@
 {
   "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
+    "outDir": "./dist",
+    "rootDir": "./src",
+    "composite": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist", "configs", "reports"],
+  "references": [
+    {
+      "path": "../../../packages/go-common"
+    }
+  ]
 }
